Refactor init to builder
