+++ Client/Application/SceneDelegate.swift
@@ -129,7 +129,8 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     /// Use this method to handle Handoff-related data or other activities.
     func scene(_ scene: UIScene, continue userActivity: NSUserActivity) {
         if AppConstants.useCoordinators {
-            // FXIOS-5985: Handle deeplinks from userActivity
+            guard let route = routeBuilder.makeRoute(userActivity: userActivity) else { return }
+            sceneCoordinator?.handle(route: route)
         } else {
             if userActivity.activityType == SiriShortcuts.activityType.openURL.rawValue {
                 browserViewController.openBlankNewTab(focusLocationField: false)
