+++ Client/Coordinators/Router/RouteBuilder.swift
@@ -3,6 +3,7 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
 import Foundation
+import CoreSpotlight
 
 struct RouteBuilder {
     private var isPrivate: () -> Bool
@@ -134,5 +135,34 @@ struct RouteBuilder {
         case .widgetTabsLargeOpenUrl:
             TelemetryWrapper.recordEvent(category: .action, method: .open, object: .largeTabsOpenUrl)
         }
+    }
+
+    func makeRoute(userActivity: NSUserActivity) -> Route? {
+        // If the user activity is a Siri shortcut to open the app, show a new search tab.
+        if userActivity.activityType == SiriShortcuts.activityType.openURL.rawValue {
+            return .search(url: nil, isPrivate: false)
+        }
+
+        // If the user activity has a webpageURL, it's a deep link or an old history item.
+        // Use the URL to create a new search tab.
+        if let url = userActivity.webpageURL {
+            return .search(url: url, isPrivate: false)
+        }
+
+        // If the user activity is a CoreSpotlight item, check its activity identifier to determine
+        // which URL to open.
+        if userActivity.activityType == CSSearchableItemActionType {
+            guard let userInfo = userActivity.userInfo,
+                  let urlString = userInfo[CSSearchableItemActivityIdentifier] as? String,
+                  let url = URL(string: urlString)
+            else {
+                return nil
+            }
+            return .search(url: url, isPrivate: false)
+        }
+
+        // If the user activity does not match any of the above criteria, return nil to indicate that
+        // the route could not be determined.
+        return nil
     }
 }
